Move textsplitter only logic to own module
